+++ Tunny.Core/Settings/Optimize.cs
@@ -1,4 +1,4 @@
-using Tunny.Core.Enum;
+using Tunny.Core.TEnum;
 using Tunny.Settings.Sampler;
 
 namespace Tunny.Core.Settings
