+++ Tunny.Core/TEnum/EndState.cs
@@ -1,4 +1,4 @@
-namespace Tunny.Core.Enum
+namespace Tunny.Core.TEnum
 {
     public enum EndState
     {
