+++ Tunny/Handler/OptimizeLoop.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.ComponentModel;
 
 using Tunny.Component.Optimizer;
-using Tunny.Core.Enum;
 using Tunny.Core.Handler;
 using Tunny.Core.Input;
 using Tunny.Core.Settings;
+using Tunny.Core.TEnum;
 using Tunny.Core.Util;
 using Tunny.Input;
 using Tunny.PostProcess;
