+++ Tunny/Input/Objective.cs
@@ -12,7 +12,7 @@ using Python.Runtime;
 using Rhino.Geometry;
 
 using Tunny.Component.Params;
-using Tunny.Core.Enum;
+using Tunny.Core.TEnum;
 using Tunny.Core.Util;
 using Tunny.Type;
