+++ Tunny/Solver/HumanInTheLoop/HumanInTheLoopBase.cs
@@ -5,7 +5,6 @@ using System.Windows.Forms;
 
 using Python.Runtime;
 
-using Tunny.Core.Handler;
 using Tunny.Core.Settings;
 using Tunny.Core.Util;
 using Tunny.UI;
