+++ Tunny/UI/OptimizeWindowTab/FileTab.cs
@@ -3,8 +3,8 @@ using System.Diagnostics;
 using System.IO;
 using System.Windows.Forms;
 
-using Tunny.Core.Enum;
 using Tunny.Core.Storage;
+using Tunny.Core.TEnum;
 using Tunny.Core.Util;
 
 namespace Tunny.UI
@@ -54,7 +54,7 @@ namespace Tunny.UI
                 string storagePath = _settings.Storage.Path;
                 if (!File.Exists(storagePath))
                 {
-                    new StorageHandler().CreateNewStorage(true, _settings.Storage);
+                    new StorageHandler().CreateNewTStorage(true, _settings.Storage);
                 }
                 UpdateStudyComboBox(storagePath);
             }
