+++ Tunny/UI/OptimizeWindowTab/OptimizeTab.cs
@@ -8,9 +8,9 @@ using Grasshopper.GUI;
 
 using Optuna.Study;
 
-using Tunny.Core.Enum;
 using Tunny.Core.Handler;
 using Tunny.Core.Storage;
+using Tunny.Core.TEnum;
 using Tunny.Core.Util;
 using Tunny.Handler;
