+++ Tunny/UI/OptimizeWindowTab/SettingsTab.cs
@@ -2,7 +2,7 @@ using System;
 
 using Serilog.Events;
 
-using Tunny.Core.Enum;
+using Tunny.Core.TEnum;
 using Tunny.Core.Util;
 using Tunny.Settings.Sampler;
